requirements.txt
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
hurry.filesize==0.9
idna==3.4
Pillow==9.4.0
praw==7.7.0
prawcore==2.3.0
requests==2.28.2
soupsieve==2.4
tqdm==4.64.1
update-checker==0.18.0
urllib3==1.26.14
websocket-client==1.5.1
